refactor: move LogSourceCard as nested component of logonboarding page
