+++ web/src/pages/LogSourceOnboarding/LogSourceOnboarding.tsx
@@ -19,7 +19,6 @@
 import React from 'react';
 import { Box, Card, FadeIn } from 'pouncejs';
 import urls from 'Source/urls';
-import LogSourceCard from 'Components/LogSourceCard';
 import s3Logo from 'Assets/s3-minimal-logo.svg';
 import sqsLogo from 'Assets/sqs-minimal-logo.svg';
 import gsuiteLogo from 'Assets/gsuite-minimal-logo.svg';
@@ -27,6 +26,7 @@ import eventBridgeLogo from 'Assets/eventbridge-minimal-logo.svg';
 import oktaLogo from 'Assets/okta-minimal-logo.svg';
 import withSEO from 'Hoc/withSEO';
 import { WizardPanel } from 'Components/Wizard';
+import LogSourceCard from './LogSourceCard';
 
 const logSourcesConfig = [
   {
